requirements/prod.txt
anyio==4.3.0
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
beautifulsoup4==4.12.3
aiohttp==3.9.3
